Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

Minor comments inline.
My main concern about this feature is that we're a little bit outside of the scope of dci modules.
I know this is a nice feature and I'm also using it but by managing this directly in the playbook (~3/4 lines of codes)

Patch-set: 3
Label: Code-Review=-1