Update patch set 7
Patch Set 7:

> No, I did this on purpose. job_id is an optional parameter

Ok I didn't see the update on the latest patch set but you forgot to update the parameter documentation to reflect id/job_id

> if the parameter is not set, we use the current job_id, like with the other modules.

There is no modules (except dci_job of course) where we are using job.list to get the job id. It's always a parameter.
It doesn't make sense to query the api to get the current job id because we already get it during the job schedule.

Patch-set: 7