sample: Add upgrade workflow
Add a sample to represent the upgrade workflow. This is done by using
dci_job with the upgrade flag set to true at the end of a classic
deployment job.
The information related to the job upgrade (components, jobs, tests)
are still stored in the job_informations variables.

You can use the job_id from the previous job_informations:
  dci_job:
    id: "{{ job_informations['id'] }}"
    upgrade: true
  register: job_informations

Or use a static value:
  dci_job:
    id: 'dd963944-abdd-4aff-bcad-14784c0c9ab9'
    upgrade: true
  register: job_informations

Change-Id: I625d4febf32a8689222476c7c286e1254f585ad0