Update patch set 4
Patch Set 4: Code-Review-1

LGTM. But, please follow the pattern used on the whole project module.params['X'] rather than a the module.params.get('X') pattern so we can be consistent. module.params will actually create all the field pass in the constructor, which mean module.params['X'] will not fail but would be set to None, so no need for the .get() (I suppose this is the reason it was here in the first place). Else +2 for me

Patch-set: 4
Label: Code-Review=-1