Refact of module code
Modules are all doing the same thing, the same way.
Yet the code is repeated in every module file. Meaning that,
when adding a new feature, this needs to be add in all N modules file.

This commit refactors the code, so we avoid repeating ourselves as much
as possible.

All modules inherits a base class that know how to do modules action.

Along this change an improvement in error management has been added

Change-Id: I431d8e2f8f0c9171a030e91a61c1f19c4730ff38