Introduce the error state
This commit introduces the error state as a jobstate.  A job would result
in error if anything wrong happens during the setup of the platform (new
and pre-run). This would mean that the job isn't usefull for
metrics/feedback as it didn't get to test the actual product.

Change-Id: I7e1659eadf3b6bbe5b81c3d60f021c0dfe8c65cd
Trello: https://trello.com/c/FK6v7k35