Update patch set 3
Patch Set 3: Code-Review-1

If we want to keep the same architecture, you'll need to move the keys action under dci_remoteci file because refresh_keys is part of the remoteci bindings.

Currently we could thought that this module is using /keys endpoint on the api which is wrong.

This could be something like:
- name: Refresh remoteci keys
  dci_remoteci:
    id: XXXXX
    refresh_keys: true

WDYT ?

Patch-set: 3
Label: Code-Review=-1