Update patch set 3
Patch Set 3: Code-Review+2

@Goneri: The issue is unrelated to the patch itself, but here the explanation of what's happening:

This issue arises when you run the test-suite twice.

What happens is that a role "Manager" is created, then deleted (soft-delete) - but not purged.

Since commit[1] we don't purge the deleted resources. So here one option would be to expand the clean_db, or simply revert to "dcictl purge" and be in full-control of what we are doing.

Why this wasn't happening earlier? Last week a patch has been merged that changed the constraint (resource, status)[2] that triggered this behavior.

We don't run the test suite twice, hence it couldn't be detected.


[1] https://github.com/redhat-cip/dci-ansible/commit/8fc454daffbd7f7871c76a52cce668729f9337e3
[2] https://softwarefactory-project.io/r/#/c/12197/

Patch-set: 3
Reviewer: Gerrit User 115 <115@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=+2