Update patch set 2
Patch Set 2:

@Goneri: The issue is unrelated to this patch.


The issue arises because the docker image for dci-control-server isn't configured to handle the new keys related feature. However it is configured in the CI as we use dci-infra-all-in-one. A patchset has been merged this morning to bring this feature to dci-control-server/Dockerfile[1]. Please rebuild the image and give it a new try.

Also there was a typo in dci-ansible[2] fixed also this morning.


[1] https://softwarefactory-project.io/r/#/c/12435/
[2] https://github.com/redhat-cip/dci-ansible/commit/542e8d1f3d46a60ae45c5c7ec166b909fbe66e04

Patch-set: 2
Label: Code-Review=0