Update patch set 6
Patch Set 6:

the pg92 client works fine, I don't want to any over-engineering that espcially because psql is just used for one call. I would actually prefer to just skip that call if you see that as a problem. I'm not sure what is the problem with the pg command and dbhost?

Patch-set: 6
Label: Code-Review=0