Update patch set 5
Patch Set 5: Code-Review-1

For me you should avoid using flags as function parameters.

It make the code harder to read and to understand. And it reveals the code is not following single responsibility principle.

So instead of:

    def post_message(self, result, output, is_skipped=False):

IMO it's better to have

    def post_skipped_message(self, result, output):

and 

    def post_message(self, result, output):


And it looks like the only thing changing is the name, so maybe add name as a parameter instead of result.

Anyway, is it possible to refactor this and use a more declarative way to describe the two functionalities ?

Patch-set: 5
Label: Code-Review=-1