do not try to create a job state before having a job_id
Before having a job_id, job states are put in a backlog and submitted
as soon as a job id is defined.

This also prevent a playbook to be blocked if it cannot access the
control server. A simple reproducer is this playbook with the dci
callback activated with an invalid https proxy configured:

- name: "Reproducer"
  hosts: localhost
  gather_facts: no
  tasks:
    - name: "Always fail"
      fail:
        msg: "blocked or not blocked?"

Change-Id: I6b2a02001945ef0330549036aa9a9b4c6738f494