Checking for actioned items in skipped loop tasks
Currently, the DCI CS shows tasks as skipped when in truth they are loop
tasks scanning a list of items in search of a specific one. For instance
a particular component type in the list of components.

This means that the task will skip most of the items, and will only take
action for the element it's looking for.

When the task completes, the last item status is used to determine the
overall status of the task so, if the last item is skipped, the task is
considered skipped when ther may be an item different than the last one
which was not skipped.

This task assesses whether a skipped loop task actually has an actioned
item and removes the skip prefix from the task file name.

Change-Id: Ibca5910568532977be3dd5ee1acc107e14625768